Added RoomsDatabase.java, and added Repository.java
In Rooms Database, followed the singleton pattern to ensure that only single instance of database is used across the entire app, also used for database builder, used addAutoMigration Strategy, to ensure the auto migration, if any column is added. Also, inside room database, get the instance of the database inside the constructor, also, initialize the historyItemDao, by using the abstract method of Database class. Also, used executor service object to ensure that database ops, will be performed on background thread. Also, used newFixedThreadPool strategy to  ensure that 4 threads or tasks can run concurrently, this will improve the responsiveness of the app. Finally, created all the functions that are defined in DAO, so that repository will be used to perform all the DAO ops.